requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pypdf
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
aiofiles
requests
pdfplumber